requirements.txt
gunicorn
Flask
psycopg2-binary
flask-cors
python-dotenv
sys